moving mocks to armhelpers and running go fmt on them again.
I'm very confused how I was faily gofmt as my editor runs it on save,
but my editor and jenkins gofmt disagree on this single file but not on
the others
